+++ src/App.js
@@ -6,7 +6,8 @@ import './App.scss';
 import HomePage from './views/pages/HomePage/HomePage';
 import QuizCollections from './views/pages/QuizCollections/QuizCollections' //for test
 
-import Quiz from './views/components/Quiz/Quiz';
+
+import QuizPage from './views/components/QuizPage/QuizPage'
 import Header from './views/components/Header/Header';
 import Footer from './views/components/Footer/Footer';
 
@@ -18,7 +19,7 @@ class App extends React.Component {
         <main>
           <Switch>
             <Route exact path='/' component={HomePage} /> 
-            <Route exact path='/big-o' component={Quiz} />
+            <Route exact path='/big-o' component={QuizPage} />
             <Route exact path='/quiz-collections' component={QuizCollections} />
           </Switch>
         </main>
